Clean up parser split
