+++ src/OneImlx.Terminal/Runtime/TerminalRequestParsedSplit.cs
@@ -1,23 +1,23 @@
 ﻿/*
-    Copyright (c) 2023 Perpetual Intelligence L.L.C. All Rights Reserved.
+    Copyright © 2019-2025 Perpetual Intelligence L.L.C. All rights reserved.
 
     For license, terms, and data policies, go to:
     https://terms.perpetualintelligence.com/articles/intro.html
 */
 
-namespace OneImlx.Terminal.Commands.Parsers
+namespace OneImlx.Terminal.Runtime
 {
     /// <summary>
     /// Represents an parsed segment split based on a token.
     /// </summary>
-    public sealed class ParsedSplit
+    public sealed class TerminalRequestParsedSplit
     {
         /// <summary>
         /// Initialize a new instance.
         /// </summary>
         /// <param name="split">The split option.</param>
         /// <param name="token">The token used to split.</param>
-        public ParsedSplit(string split, string? token)
+        public TerminalRequestParsedSplit(string split, string? token)
         {
             Split = split;
             Token = token;
